# packages/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ContaOut": {
        "properties": {
          "conta": {
            "$ref": "#/components/schemas/Conta"
          }
        },
        "required": [
          "conta"
        ],
        "title": "ContaOut",
        "type": "object"
      },
      "ContaStatus": {
        "enum": [
          "planejada",
          "ativa",
          "pausada",
          "encerrada"
        ],
        "title": "ContaStatus",
        "type": "string"
      },
      "CreateContaIn": {
        "properties": {
          "handle": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "notes": {
            "default": "",
            "maxLength": 500,
            "title": "Notes",
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "platformName": {
            "default": "",
            "maxLength": 40,
            "title": "Platformname",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ContaStatus",
            "default": "planejada"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 500,
                "pattern": "^https?://\\S+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "platform"
        ],
        "title": "CreateContaIn",
        "type": "object"
      },
      "CreatePerfilIn": {
        "properties": {
          "bio": {
            "default": "",
            "maxLength": 2000,
            "title": "Bio",
            "type": "string"
          },
          "language": {
            "default": "pt-BR",
            "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
            "title": "Language",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "niche": {
            "default": "",
            "maxLength": 200,
            "title": "Niche",
            "type": "string"
          },
          "slug": {
            "maxLength": 60,
            "minLength": 2,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PerfilStatus",
            "default": "em_preparacao"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "CreatePerfilIn",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "PerfilDetail": {
        "properties": {
          "contas": {
            "items": {
              "$ref": "#/components/schemas/Conta"
            },
            "title": "Contas",
            "type": "array"
          },
          "perfil": {
            "$ref": "#/components/schemas/Perfil"
          }
        },
        "required": [
          "perfil",
          "contas"
        ],
        "title": "PerfilDetail",
        "type": "object"
      },
      "PerfilOut": {
        "properties": {
          "perfil": {
            "$ref": "#/components/schemas/Perfil"
          }
        },
        "required": [
          "perfil"
        ],
        "title": "PerfilOut",
        "type": "object"
      },
      "PerfilStatus": {
        "enum": [
          "em_preparacao",
          "ativo",
          "pausado"
        ],
        "title": "PerfilStatus",
        "type": "string"
      },
      "PerfisList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Perfil"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PerfisList",
        "type": "object"
      },
      "Platform": {
        "enum": [
          "tiktok",
          "youtube",
          "instagram",
          "kwai",
          "facebook",
          "x",
          "outra"
        ],
        "title": "Platform",
        "type": "string"
      },
...
      "RevertIn": {
        "properties": {
          "toVersion": {
            "minimum": 1.0,
            "title": "Toversion",
            "type": "integer"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "toVersion"
        ],
        "title": "RevertIn",
        "type": "object"
      },
...
      "UpdateContaIn": {
        "additionalProperties": false,
        "properties": {
          "handle": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "platformName": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platformname"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContaStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 500,
                "pattern": "^https?://\\S+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "UpdateContaIn",
        "type": "object"
      },
      "UpdatePerfilIn": {
        "additionalProperties": false,
        "properties": {
          "bio": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "language": {
            "anyOf": [
              {
                "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "niche": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerfilStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "UpdatePerfilIn",
        "type": "object"
      },
...
      "VersionIn": {
        "properties": {
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "VersionIn",
        "type": "object"
      },
      "VersionsList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Version"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "VersionsList",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SociMan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/contas/{conta_id}": {
      "patch": {
        "operationId": "contas_update",
        "parameters": [
          {
            "in": "path",
            "name": "conta_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conta Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Update Conta"
      }
    },
    "/api/contas/{conta_id}/archive": {
      "post": {
        "operationId": "contas_archive",
        "parameters": [
          {
            "in": "path",
            "name": "conta_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conta Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Archive Conta"
      }
    },
    "/api/contas/{conta_id}/restore": {
      "post": {
        "operationId": "contas_restore",
        "parameters": [
          {
            "in": "path",
            "name": "conta_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conta Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Restore Conta"
      }
    },
    "/api/contas/{conta_id}/revert": {
      "post": {
        "operationId": "contas_revert",
        "parameters": [
          {
            "in": "path",
            "name": "conta_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conta Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Revert Conta"
      }
    },
    "/api/contas/{conta_id}/versions": {
      "get": {
        "operationId": "contas_versions",
        "parameters": [
          {
            "in": "path",
            "name": "conta_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conta Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionsList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dados inválidos (validation_error)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "List Versions"
      }
    },
...
    "/api/perfis": {
      "get": {
        "operationId": "perfis_list",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PerfilStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfisList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dados inválidos (validation_error)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "List Perfis"
      },
      "post": {
        "operationId": "perfis_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePerfilIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Create Perfil"
      }
    },
...
    "/api/perfis/{perfil_id}": {
      "get": {
        "operationId": "perfis_get",
        "parameters": [
          {
            "in": "path",
            "name": "perfil_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Perfil Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dados inválidos (validation_error)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Get Perfil"
      },
      "patch": {
        "operationId": "perfis_update",
        "parameters": [
          {
            "in": "path",
            "name": "perfil_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Perfil Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePerfilIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Update Perfil"
      }
    },
...
    "/api/perfis/{perfil_id}/contas": {
      "post": {
        "operationId": "contas_create",
        "parameters": [
          {
            "in": "path",
            "name": "perfil_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Perfil Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Create Conta"
      }
    },
...
  }
}